Added css module related config to Test.js
Prior to this fix running the tests for components created w/ css
modules and css preprocessors (less, s(a|c)ss, stylus) would fail
because the loader was configured to be the null-loader for these file
types. As a result, the style object required by css modules would
resolve to `undefined`, which in turn lets tests on these components
fail with a 'CSS Module is undefined' error.

Additionally:
- fixed eslint warnings on Base.js
- moved webpack from devDependencies to dependencies